SSL_PRETRAIN_MOCO/requirements.txt
os
pandas
torch
torchvision
opencv-python
scikit-learn
scikit-image
matplotlib
tqdm
numpy
imageio
Pillow